Handle defined protocol exceptions correctly
This PR fixes issues around documented protocol exceptions.
Previously these were grouped into the general "error" bucket, but
this is incorrect. If the handler threw a known exception this was:

* logged as an error
* counted as server RPC error
* counted as an individual RPC error
* tracked as an unknown error

This PR reverse all those. All error handlers in the codebase
have been updated to only handle unexpected errors. This is done by
using ruby's const_get methods to introspect the generated intermediate
result classes. The result classes contain the expected exceptions.
These exceptions are passed along with the internal RPC struct.

This makes metrics more clear as well. When a documented exception case
occurs, this is added to ".exception" key. For example "rpc.exception"
for server level metrics, and "rpc.getItems.exception" for the getItems
RPC.

The main entry point to this library must change as well to accommodate
this feature. Previously, the ThriftServer.wrap call took a Processor
class. Now it requires the service constant (class/namespace containing
the Processor & various xxx_result classs). Example change is:

ThriftServer.build(MyService::Processor) to ThriftServer.build(MyService).

This change also makes the API read a bit more intention revealing.
The processor is just an intermediate object for the most part, so it's
nice if removed from the public API.