Add error handling for remote_address stats
This handles an issue where the remote address cannot be resolved for various
reasons. Instead of breaking the server, we'll fall back to the default
behavior of the Apache Thrift implementation of the threaded server, where no
logging is performed.

This may not be the best technical solution. It could be improved by throwing
the exception upwards for the server implementor to handle.